Refactoring credentials into foursquare-credentials
