+++ foursquare/foursquare.js
@@ -32,12 +32,11 @@ module.exports = function(RED) {
         );
     }
     
-    
-    function SwarmNode(n) {
+    function FoursquareNode(n) {
         RED.nodes.createNode(this,n);
     }
     
-    RED.nodes.registerType("swarm-credentials", SwarmNode, {
+    RED.nodes.registerType("foursquare-credentials", FoursquareNode, {
         credentials: {
             displayname: {type: "text"},
             clientid: {type: "password"},
@@ -45,113 +44,8 @@ module.exports = function(RED) {
             accesstoken: {type: "password"}
         }
     });
-
-    /**
-     * Swarm input node - will return the most recent check-in since
-     * the node has been initialized. The node will only populate msg.payload 
-     * with the JSON of the check-in if a new check-in has been made within 
-     * the polling interval.
-     */
-    function SwarmInNode(n) {
-        RED.nodes.createNode(this, n);
-        var credentials = RED.nodes.getCredentials(n.swarm);
-        var node = this;
-        var credentialsOk = checkCredentials(node, credentials);
-        if (credentialsOk) {
-            var repeat = 900000; // 15 mins
-            var now = Math.floor(((new Date()).getTime())/1000); // time now in seconds since epoch
-            var afterTimestamp = now;
-            var lastcheckin = null;
-            
-            var interval = setInterval(function() {
-                node.emit("input", {});
-            }, repeat );
-            
-            this.on("input", function(msg) {
-                getCheckinsAfterTimestamp(node, "self", afterTimestamp, credentials, msg, function(msg) {
-                    var latestcheckin = JSON.stringify(msg);
-                    if (latestcheckin != lastcheckin) {
-                        lastcheckin = latestcheckin;
-                        afterTimestamp = msg.payload.createdAt;
-                        node.send(msg);
-                    }
-                }); 
-            });                       
-
-            this.on("close", function() {
-                if (interval != null) {
-                    clearInterval(interval);
-                }          
-            });
-
-        }
-    } 
     
-    RED.nodes.registerType("swarm in", SwarmInNode);
-
-    /**
-     * Swarm query node - will return the most recent check-in since
-     * the node has been initialized. If a check-in exists the node will always return 
-     * the most recent even if no new check-ins have happened since the previous query.
-     * The node only populates msg.payload when a check-in is found.  
-     */
-    function SwarmQueryNode(n) {
-        RED.nodes.createNode(this, n);
-        var node = this;
-        this.request = n.request || "get-most-recent-checkin";
-        var credentials = RED.nodes.getCredentials(n.swarm);
-        var credentialsOk = checkCredentials(node, credentials);
-        if (credentialsOk) {
-          var now = Math.floor(((new Date()).getTime())/1000); // time now in seconds since epoch (rounded down)
-          var afterTimestamp = now;
-          
-          this.on("input", function(msg) {
-              if (node.request === "get-most-recent-checkin") {
-                  getCheckinsAfterTimestamp(node, "self", afterTimestamp, credentials, msg, function(msg) {
-                      afterTimestamp = msg.payload.createdAt - 2;
-                      node.send(msg);
-                  });                                  
-              }
-          });            
-        }
-    } 
-    
-    RED.nodes.registerType("swarm", SwarmQueryNode); 
-    
-    function checkCredentials(node, credentials) {       
-        if (credentials && credentials.clientid && credentials.clientsecret && credentials.accesstoken) {
-           return true;
-        } else {
-            node.error("problem with credentials being set: " + credentials + ", ");
-            node.status({fill:"red",shape:"ring",text:"failed"});      
-            return false;
-        }
-    }
-    
-    function getCheckinsAfterTimestamp(node, userid, afterTimestamp, credentials, msg, callback) {
-        var apiUrl = "https://api.foursquare.com/v2/users/" + userid + "/checkins?oauth_token=" + credentials.accesstoken  + "&v=20141016&afterTimestamp=" + afterTimestamp+"&sort=newestfirst";
-        request.get(apiUrl,function(err, httpResponse, body) {
-            if (err) {
-                node.error(err.toString());
-                node.status({fill:"red",shape:"ring",text:"failed"});
-            } else {
-                var result = JSON.parse(body);
-                if (result.meta.code != 200) {
-                    node.error("Error code: " + result.meta.code + ", errorDetail: " + result.meta.errorDetail);
-                    node.status({fill:"red",shape:"ring",text:"failed"});
-                } else {
-                    if (result.response.checkins.items.length !== 0) {
-                      var latest = result.response.checkins.items[0];
-                      msg.payload = {};
-                      msg.payload = latest;
-                      callback(msg);
-                  }                    
-                }
-            }
-        });              
-    }
-    
-    RED.httpAdmin.get('/swarm-credentials/auth', function(req, res){
+    RED.httpAdmin.get('/foursquare-credentials/auth', function(req, res){
         if (!req.query.clientid || !req.query.clientsecret || !req.query.id || !req.query.callback) {
             return res.status(400).send('ERROR: request does not contain the required parameters');
         }
@@ -175,7 +69,7 @@ module.exports = function(RED) {
         res.redirect(url);
     });
 
-    RED.httpAdmin.get('/swarm-credentials/auth/callback', function(req, res){
+    RED.httpAdmin.get('/foursquare-credentials/auth/callback', function(req, res){
         if (req.query.error) {
             return res.send("ERROR: " + req.query.error + ": " + req.query.error_description);
         }
